pyproject.toml
[tool.poetry]
name = "obsidian_boy"
version = "0.1.0"
description = "An Obsidian AI assistant"
authors = ["Olaf Geibig <olaf@pobox.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
langchain = "^0.2.16"
langgraph = "^0.2.16"
langchain-openai = "^0.1.23"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
langchain-groq = "^0.1.9"
langchain-anthropic = "^0.1.23"
beautifulsoup4 = "^4.12.3"
markdownify = "^0.13.1"
arize-phoenix-otel = "^0.4.1"
openinference-instrumentation-openai = "^0.1.14"
openinference-instrumentation-langchain = "^0.1.28"
crewai = "^0.63.2"
openinference-instrumentation-crewai = "^0.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-integration = "^0.2.3"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.scripts]
mycov = "pytest --without-integration --cov=obsidian_boy"
mytest = "pytest --without-integration"